# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.CreateApplicationDTO:
    properties:
      candidate_id:
        minimum: 1
        type: integer
      company_id:
        minimum: 1
        type: integer
      job_id:
        minimum: 1
        type: integer
      rating:
        maximum: 5
        minimum: 1
        type: integer
      stage:
        enum:
        - applied
        - screening
        - technical
        - offer
        - hired
        - rejected
        type: string
    required:
    - candidate_id
    - company_id
    - job_id
    - stage
    type: object
...
  dtos.CreateCompanyWithAdminDTO:
    properties:
      admin_email:
        type: string
      admin_first_name:
        type: string
      admin_last_name:
        type: string
      admin_password:
        minLength: 8
        type: string
      company_name:
        type: string
      company_slug:
        type: string
      plan_slug:
        description: Opcional, default "free"
        type: string
    required:
    - admin_email
    - admin_first_name
    - admin_last_name
    - admin_password
    - company_name
    - company_slug
    type: object
...
host: localhost:8001
info:
  contact:
    email: support@dvra.com
    name: API Support
  description: API para sistema de reclutamiento y gestión de candidatos
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DVRA API
  version: 1.2.0
paths:
  /admin/analytics:
    get:
      consumes:
      - application/json
      description: Obtiene estadísticas globales de la plataforma
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Analíticas globales (SuperAdmin)
      tags:
      - SuperAdmin
  /admin/companies:
    get:
      consumes:
      - application/json
      description: Obtiene todas las empresas con paginación y filtros
      parameters:
      - default: 1
        description: Página
        in: query
        name: page
        type: integer
      - default: 20
        description: Límite
        in: query
        name: limit
        type: integer
      - description: Búsqueda
        in: query
        name: search
        type: string
      - description: Filtrar por plan
        in: query
        name: plan_tier
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar todas las empresas (SuperAdmin)
      tags:
      - SuperAdmin
    post:
      consumes:
      - application/json
      description: Crea una nueva empresa y su usuario administrador
      parameters:
      - description: Datos de empresa y admin
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateCompanyWithAdminDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear empresa con admin (SuperAdmin)
      tags:
      - SuperAdmin
...
  /applications:
    get:
      consumes:
      - application/json
      description: Obtiene postulaciones (todas si es SuperAdmin, de la empresa si
        es usuario normal)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar postulaciones
      tags:
      - Applications
    post:
      consumes:
      - application/json
      description: Crea una nueva postulación a un empleo
      parameters:
      - description: Datos de la postulación
        in: body
        name: application
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateApplicationDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear postulación
      tags:
      - Applications
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"